requirements.txt
Flask>=2.2
Flask-Session>=0.4
Flask-Login>=0.5
Flask-Mail>=0.8
Flask-MySQL>=1.0
Flask-Bcrypt>=1.0
pusher>=2.1
py3-validate-email>=1.0; python_version > '3.1'
validate-email>=1.0; python_version <= '3.1'